refactor: content pkg update(1/2)
