+++ CHANGELOG.md
@@ -16,8 +16,8 @@ The versions coincide with releases on pip. Only major versions will be released
 ## [0.0.x](https://github.com/oras-project/oras-py/tree/main) (0.0.x)
  - restructure the copy engine into a layered call stack `provider -> copy -> content` (0.2.43)
    - the DAG copy algorithm (`copy_graph`, `StatusTracker`, `successors`) lives in `oras.copy` (`oras.copy.graph`, `oras.copy.tracker`), removing the `content <-> copy` circular import
-   - the storage/target `Protocol` contract moved to `oras.content.storage`, the in-memory store + read-through cache (`MemoryStorage`, `CacheProxy`, `FetcherFunc`) to `oras.content.memory`, and the concrete adapters `RegistryTarget`/`LayoutTarget` to `oras.content.registry`/`oras.content.layout`
-   - **backward incompatible:** import `Target`/`ReadOnlyTarget`/`Storage`/`ReadOnlyStorage`/`ReferencePusher`/`ReferenceFetcher`/`Mounter` from `oras.content.storage` (also re-exported from `oras.copy`) and `MemoryStorage`/`CacheProxy`/`FetcherFunc` from `oras.content.memory`; `RegistryTarget` moved from `oras.provider` to `oras.content.registry`; `LayoutTarget` moved from `oras.layout.layout` to `oras.content.layout`
+   - the storage/target `Protocol` contract plus the general-purpose `CacheProxy` cache and `FetcherFunc` adapter moved to `oras.content.storage`, the in-memory store (`MemoryStorage`) to `oras.content.memory`, and the concrete adapters `RegistryTarget`/`LayoutTarget` to `oras.content.registry`/`oras.content.layout`
+   - **backward incompatible:** import `Target`/`ReadOnlyTarget`/`Storage`/`ReadOnlyStorage`/`ReferencePusher`/`ReferenceFetcher`/`Mounter` (also re-exported from `oras.copy`) and `CacheProxy`/`FetcherFunc` from `oras.content.storage`, and `MemoryStorage` from `oras.content.memory`; `RegistryTarget` moved from `oras.provider` to `oras.content.registry`; `LayoutTarget` moved from `oras.layout.layout` to `oras.content.layout`
    - the descriptor primitives (`is_manifest`, `is_foreign_layer`, `descriptor_key`, `descriptors_equal`, `remove_foreign_layers`) moved from `oras.content.descriptor` to `oras.copy.descriptor`, and `oras.content.registry` no longer depends on the descriptor module — **backward incompatible:** import these helpers from `oras.copy.descriptor`
    - the unused `Fetcher`/`Pusher`/`Resolver`/`Tagger`/`TagResolver` protocols were removed
  - add Layout `copy` for pull_from_registry capability (0.2.42)
