+++ oras/content/__init__.py
@@ -1,9 +1,10 @@
 """
 OCI content layer for the copy engine.
 
-Holds the storage/target Protocol contract (:mod:`oras.content.storage`) and
-the in-memory store + read-through cache (:mod:`oras.content.memory`) shared by
-the copy algorithm and the concrete target adapters.
+Holds the storage/target Protocol contract and its general-purpose helpers
+(read-through cache, callable-fetcher adapter) in :mod:`oras.content.storage`,
+and the in-memory store in :mod:`oras.content.memory`, shared by the copy
+algorithm and the concrete target adapters.
 """
 
 __author__ = "The ORAS Authors"
