+++ oras/content/storage.py
@@ -2,10 +2,12 @@
 Storage and target interfaces for the copy engine.
 
 Defines the Protocol classes (the "from/to" storage + target contract) that
-the copy engine checks against. Concrete realizations live in
-:mod:`oras.content.memory` (in-memory store + read-through cache) and in the
-target adapters :mod:`oras.content.registry` and :mod:`oras.content.layout`; the
-copy algorithm in :mod:`oras.copy` consumes these protocols.
+the copy engine checks against, plus the general-purpose helpers built on
+them: :class:`FetcherFunc` (callable-to-fetcher adapter) and
+:class:`CacheProxy` (read-through metadata cache). Other concrete
+realizations live in :mod:`oras.content.memory` (in-memory store) and in the
+target adapters :mod:`oras.content.registry` and :mod:`oras.content.layout`;
+the copy algorithm in :mod:`oras.copy` consumes these protocols.
 
 These mirror the relevant subset of oras-go's content package interfaces.
 Only the protocols the engine actually uses are defined here:
@@ -23,6 +25,8 @@ __author__ = "The ORAS Authors"
 __copyright__ = "Copyright The ORAS Authors."
 __license__ = "Apache-2.0"
 
+import io
+import threading
 from typing import BinaryIO, Callable, Protocol, Tuple, runtime_checkable
 
 from oras.types import Descriptor
@@ -106,3 +110,76 @@ class Mounter(Protocol):
         from_repo: str,
         get_content: Callable[[], BinaryIO],
     ) -> None: ...
+
+
+# ---------------------------------------------------------------------------
+# FetcherFunc: adapter from a plain callable to a fetcher
+# ---------------------------------------------------------------------------
+
+
+class FetcherFunc:
+    """Wraps a callable as a Fetcher."""
+
+    def __init__(self, fn: Callable[[Descriptor], BinaryIO]):
+        self._fn = fn
+
+    def fetch(self, desc: Descriptor) -> BinaryIO:
+        return self._fn(desc)
+
+
+# ---------------------------------------------------------------------------
+# CacheProxy: read-through caching proxy built on the storage protocols
+# ---------------------------------------------------------------------------
+
+
+class CacheProxy:
+    """
+    Caching proxy that wraps a ReadOnlyStorage with a local Storage cache.
+
+    Non-leaf nodes (manifests, indexes) are cached in memory for reuse
+    during graph traversal. The cache has a byte size limit to prevent
+    unbounded memory growth.
+
+    Matches oras-go's internal/cas.Proxy.
+    """
+
+    def __init__(self, base: ReadOnlyStorage, cache: Storage, max_bytes: int):
+        self.base = base
+        self.cache = cache
+        self.max_bytes = max_bytes
+        self.stop_caching: bool = False
+        self._cached_bytes: int = 0
+        self._lock = threading.Lock()
+
+    def exists(self, desc: Descriptor) -> bool:
+        if self.cache.exists(desc):
+            return True
+        return self.base.exists(desc)
+
+    def fetch(self, desc: Descriptor) -> BinaryIO:
+        # Try cache first (single call avoids TOCTOU)
+        try:
+            return self.cache.fetch(desc)
+        except FileNotFoundError:
+            pass
+
+        # Fetch from base
+        stream = self.base.fetch(desc)
+
+        if self.stop_caching:
+            return stream
+
+        size = desc.get("size", 0)
+        with self._lock:
+            if self._cached_bytes + size > self.max_bytes:
+                return stream
+            self._cached_bytes += size
+
+        # Read, cache, and return
+        data = stream.read()
+        self.cache.push(desc, io.BytesIO(data))
+        return io.BytesIO(data)
+
+    def push(self, desc: Descriptor, content: BinaryIO) -> None:
+        """Push delegates to the base (not the cache)."""
+        raise NotImplementedError("CacheProxy is read-only; push to base directly")
